feat/Abstract_network_model_structure
